+++ compiler-core/src/language_server/tests/hover.rs
@@ -1,0 +1,211 @@
+use lsp_types::{
+    Hover, HoverContents, HoverParams, MarkedString, Position, Range, TextDocumentIdentifier,
+    TextDocumentPositionParams, Url,
+};
+
+use super::*;
+
+fn positioned_hover(src: &str, position: Position) -> Option<Hover> {
+    let io = LanguageServerTestIO::new();
+    let mut engine = setup_engine(&io);
+
+    _ = io.src_module("app", src);
+    let response = engine.compile_please();
+    assert!(response.result.is_ok());
+
+    let path = Utf8PathBuf::from(if cfg!(target_family = "windows") {
+        r#"\\?\C:\src\app.gleam"#
+    } else {
+        "/src/app.gleam"
+    });
+
+    let url = Url::from_file_path(path).unwrap();
+
+    let params = HoverParams {
+        text_document_position_params: TextDocumentPositionParams::new(
+            TextDocumentIdentifier::new(url),
+            position,
+        ),
+        work_done_progress_params: Default::default(),
+    };
+    let response = engine.hover(params);
+
+    response.result.unwrap()
+}
+
+#[test]
+fn hover_function_definition() {
+    let code = "
+fn add_2(x) {
+  x + 2
+}
+";
+
+    assert_eq!(
+        positioned_hover(&code, Position::new(1, 3)),
+        Some(Hover {
+            contents: HoverContents::Scalar(MarkedString::String(
+                "```gleam
+fn(Int) -> Int
+```
+"
+                .to_string()
+            )),
+            range: Some(Range {
+                start: Position {
+                    line: 1,
+                    character: 0,
+                },
+                end: Position {
+                    line: 1,
+                    character: 11,
+                },
+            },),
+        })
+    );
+}
+
+#[test]
+fn hover_function_definition_with_docs() {
+    let code = "
+/// Exciting documentation
+/// Maybe even multiple lines
+fn append(x, y) {
+  x <> y
+}
+";
+
+    assert_eq!(
+        positioned_hover(&code, Position::new(3, 3)),
+        Some(Hover {
+            contents: HoverContents::Scalar(MarkedString::String(
+                "```gleam
+fn(String, String) -> String
+```
+ Exciting documentation
+ Maybe even multiple lines
+"
+                .to_string()
+            )),
+            range: Some(Range {
+                start: Position {
+                    line: 3,
+                    character: 0,
+                },
+                end: Position {
+                    line: 3,
+                    character: 15,
+                },
+            },),
+        })
+    );
+}
+
+#[test]
+fn hover_function_argument() {
+    let code = "
+/// Exciting documentation
+/// Maybe even multiple lines
+fn append(x, y) {
+  x <> y
+}
+";
+
+    assert_eq!(
+        positioned_hover(&code, Position::new(3, 10)),
+        Some(Hover {
+            contents: HoverContents::Scalar(MarkedString::String(
+                "```gleam\nString\n```".to_string()
+            )),
+            range: Some(Range {
+                start: Position {
+                    line: 3,
+                    character: 10,
+                },
+                end: Position {
+                    line: 3,
+                    character: 11,
+                },
+            },),
+        })
+    );
+}
+
+#[test]
+fn hover_function_body() {
+    let code = "
+/// Exciting documentation
+/// Maybe even multiple lines
+fn append(x, y) {
+  x <> y
+}
+";
+
+    assert_eq!(positioned_hover(&code, Position::new(4, 1)), None);
+}
+
+#[test]
+fn hover_expressions_in_function_body() {
+    let code = "
+fn append(x, y) {
+  x <> y
+}
+";
+
+    assert_eq!(
+        positioned_hover(&code, Position::new(2, 2)),
+        Some(Hover {
+            contents: HoverContents::Scalar(MarkedString::String(
+                "```gleam
+String
+```
+A locally defined variable."
+                    .to_string()
+            )),
+            range: Some(Range {
+                start: Position {
+                    line: 2,
+                    character: 2
+                },
+                end: Position {
+                    line: 2,
+                    character: 3
+                }
+            }),
+        })
+    );
+}
+
+#[test]
+fn hover_module_constant() {
+    let code = "
+/// Exciting documentation
+/// Maybe even multiple lines
+const one = 1
+";
+
+    assert_eq!(
+        positioned_hover(&code, Position::new(3, 6)),
+        Some(Hover {
+            contents: HoverContents::Scalar(MarkedString::String(
+                "```gleam
+Int
+```
+ Exciting documentation
+ Maybe even multiple lines
+"
+                .to_string()
+            )),
+            range: Some(Range {
+                start: Position {
+                    line: 3,
+                    character: 6
+                },
+                end: Position {
+                    line: 3,
+                    character: 9
+                },
+            }),
+        })
+    );
+}